+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the `ci-warehouse` library will be documented in this change log file.
 
+## v1.0.2-alpha - 2018-12-03
+
+### Added
+- `Warehouse` exception classes
+- Validation checks for repository classes and configuration entries
+
 ## v1.0.1-alpha - 2018-12-03
 
 ### Added
@@ -9,7 +15,7 @@ All notable changes to the `ci-warehouse` library will be documented in this cha
 - Post install script to publish the config file
 
 ### Modified
-- Cleaned up the load `Warehouse` load method
+- Cleaned up the `Warehouse` load() method
 
 ## v1.0.0-alpha - 2018-12-02